Enable filtering of store products
Products on the store can now be filtered by category so that only
products of a certain category would be displayed. These filters are
accessed through the previously non-functional category links on the
navbar.

The order of categories on the navbar was also adjusted to fit the order
of display on the store page, maintaining UI consistency.