Refactor layout panels
This change simplifies the panel logic in the layout by consolidating
the functionality separate Flyout and MobileMenu components into the
layout component without the indirection of a separate component.

The panels can now adapts its style and behavior based on screen size
and panel type, reducing code duplication and complexity.

Associated CSS modules and utility hooks for the old components are
removed, and new tests are added for Layout component to ensure correct
behavior.